[BEAM-5428][backport] Implement cross-bundle user state caching in the Python SDK
This caches user state in the Python SDK using a LRU cache with a bounded number
of items. Cache entries are guarded by cache tokens which are supplied by the
Runner in every ProcessBundleRequest. Cache entries are stored throughout
bundles and are evicted either by the least-recently-used strategy or when a new
cache entry for the same state key is stored.

The caching layer sits in front of the state retrieval layer which means that
requests are going to the cache first, if a valid cache token exists and the
cache contains an entry for a state key. Upon first reading state for a given
state key, the caching layer materializes the state and stores it in the
cache.

With the recent changes to the cache token handling in the Proto, append or
clear requests are executed non-blocking because the cache tokens provided for a
bundle remain valid for the life time of a bundle.